[ Cli tools ] Fix uaserver and tests
This commits adds a server fixture running in a dedicated thread to handle OPC-UA
requests properly (the existing server fixture accept TCP connections
but are stuck at hello).

We then use this fixture for the CLI tests.

Also fixing the uaserver cli, its event_loop was stuck in a not async
sleep preventing it from handling requests.