Support variables with properties in events and extend file generators for Alarms and Conditions (#415)
* Complete client alarms and conditions example

Complete client alarms and conditions example for usage with OPCFoundation UA-.NETStandard-Samples Quickstart AlarmConditionServer

* Add variables to generated event filter and event object

Event types can not only contain properties, but also variables. For example the ConditionType, which is the foundation of Alarms and Conditions, contains variables. Add the variables to the generated event filter and event object.

* Changed schemas for Fix localtime field of events

* changed schemas to autogenerate event_objects.py with conditions and alarms

* expand the event model with variables which are able to have properties as children
also supports custom events
now we are able to generate alarms and conditions

* Some refactoring and commenting in events and generate_model_event

* Newly generated event_objects with Alarms and Conditions Types and Variables

Co-authored-by: mar-ar <marco.arnold@hermle.de>
Co-authored-by: oroulet <oroulet@users.noreply.github.com>