Fix `MonitoredItemServer._is_data_changed`, with as trigger `ua.DataChangeTrigger.StatusValueTimestamp` is used.
Currently with  `ua.DataChangeTrigger.StatusValueTimestamp` it only triggers when the SourceTimeStamp or the StatusCodes changes. Not on value updates.

According to Part 4 7.17.2 DataChangeFilter when the trigger is `ua.DataChangeTrigger.StatusValueTimestamp` is should trigger if either StatusCode, value or SourceTimeStamp changes.

No tests where available for monitored items, test added for this situation.