Add and demonstrate convenient helper for self-signed certificates
With PR #1368 some functions are added to create and request certificates.
Based on those functions a new helper method `setup_self_signed_certificate` is created.
This should make it easy to add basic support to clients and servers for generating automatically certificates.

If can generated a private key and self-signed certificate when not already present.
When the data range of the certificate is no longer valid it generates a new certificate.

Both the examples for client and server use with encryption are updated to demonstrate the use.