Adds a certificate validation hook to the client and server
With the hook you can register callback method with the following type `CertificateValidatorMethod`:
```
Callable[[x509.Certificate, ApplicationDescription], Awaitable[None]]
```

The method should raise an `ServiceError` exception when the certificate is invalid.
The makes it possible to provide your own logic for validation.

An default implementation of such certificate validation hook is also provided by `ucrypto.validator.CertificateValidator`.

`CertificateValidator` supports checking:
- Uri
- Timerange
- Key Usage
- Extended Key Usage
- Trust
- Revoked

For the last two checks a instance of a `TrustStore` needs to be provided to `CertificateValidator`.

Which checks must be performed can be configured with `CertificateValidatorOptions`. Those are provide to the constructor of `CertificateValidator` and can be changed later with `CertificateValidator.set_validate_options`.

Examples of client and server with encryption are changed to demonstrate the use of the validator.

Basic testing of the validator is present.